Add comments for every class
I added comments to every class and method as per the java coding
standard
