--- src/main/java/Task/Deadline.java
@@ -1,15 +1,0 @@
-package Task;
-public class Deadline extends Task {
-
-    protected String by;
-
-    public Deadline(String description, String by) {
-        super(description);
-        this.by = by;
-    }
-
-    @Override
-    public String toString() {
-        return "[D]" + super.toString() + " (by: " + by + ")";
-    }
-}
