--- src/main/java/Task/Todo.java
@@ -1,11 +1,0 @@
-package Task;
-public class Todo extends Task{
-    public Todo(String text) {
-        super(text);
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
