+++ src/main/java/msg/StdMsgs.java
@@ -1,7 +1,9 @@
-package Msg;
-public enum Std_msgs {
-    // collection of std msgs to be used
-    WELCOME(new Msg("Greetings Sir/Mdm, I am Sir Duke, at your service\n" +
+package msg;
+/** This enum class represents a collection of standard messages or responses Sir Duke will say.
+ * His logo is also stored here
+ */
+public enum StdMsgs {
+    WELCOME(new Msg("Greetings Sir/Mdm, I am Sir sirduke, at your service\n" +
             "How can I help?")),
     BYE(new Msg("Goodbye Sir/Mdm")),
     LOGO(new Msg("Greetings from\n" +
@@ -15,11 +17,16 @@ public enum Std_msgs {
     MARK(new Msg("Nice! I've marked this task as done:", false)),
     UNMARK(new Msg("OK, I've marked this task as not done yet:", false));
 
-    private final Msg msg;
-    Std_msgs(Msg msg){
+
+    /** Every Std_Msgs is of the message class */
+    private Msg msg;
+    StdMsgs(Msg msg){
         this.msg = msg;
     }
 
+    /**
+     * Returns the message in StdMsgs is a message
+     */
     @Override
     public String toString() {
         return this.msg.toString();
