+++ text-ui-test/EXPECTED.TXT
@@ -8,7 +8,7 @@ Greetings from
 
 ____________________________________________________________
 
-Greetings Sir/Mdm, I am Sir Duke, at your service
+Greetings Sir/Mdm, I am Sir sirduke, at your service
 How can I help?
 ____________________________________________________________
